Create Assignment 3 = CricketRunRateCalculator
For an ongoing cricket match, data about overs bowled, current run rate and target runs is
known. The captain needs help in computing the required run rate to win the match.
Write a program to compute the required run rate. Program should take overs bowled, current
run rate and the target as input. Consider 50 over match.
Calculation:
Input: overs bowled = 10, current run rate = 4.2, target = 250
Required run rate = (250 – (4.2*10)) / 40 = 5.2
Input:
The first input is an integer which corresponds to the number of overs bowled so far. The
second input is a float which corresponds to the current run rate. The third input is an integer
which corresponds to the target score.
Sample
Enter the number of overs bowled so far
10
Enter the current run rate
4.2
Enter the target score
250