Create Assignment 4 = SpecialSequenceGenerator
Assignment 4 – Special Sequence
In a special sequence of numbers, 2, 1 and 3 are the 1st three numbers. All other numbers in
the sequence are generated by the sum of their 3 most recent predecessors.
Write a program to generate this special sequence where the last number in the sequence is
less than or equal to the input number ‘n’
Input Format:
Input consists of a single integer which corresponds to n.
Output Format:
Output consists of the Sequence, separated by a single space. There are no leading or
trailing spaces in the output.
Sample Input
20
Sample Output
2 1 3 6 10 19