Create Assignment 6 = KaprekarNumber
Kaprekar number
Consider an n-digit number k. Square it and add the right n digits to the left n or n-1 digits. If
the resultant sum is k, then k is called a Kaprekar number.
For example, 9 is a Kaprekar number since
92 = 81
& 8+1=9.
297 is a Kaprekar number since
2972 = 88209
& 88+209 = 297.
Write a program to find whether a given number is a Kaprekar number or not.

Input Format:
Input consists of a single integer.
Output Format:
Refer sample output for details.

Sample Input 1:
9
Sample Output 1:
Kaprekar Number

Sample Input 2:
101
Sample Output 2:
Not A Kaprekar Number